#60 some code to enable log manipulation
