+++ cache/store/store.go
@@ -5,7 +5,7 @@ import (
 	"github.com/mageddo/dns-proxy-server/cache"
 )
 
-var c cache.Cache;
+var c cache.Cache
 func init(){
 	c = lru.New(43690); // about 1 MB considering HostnameVo struct
 }
