+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
   # Base Image to build project
   gocompiler-docker-dns-proxy:
     image: golang:1.9
-    container_name: gocompiler
+    container_name: dps-gocompiler
     working_dir: /app/src/github.com/mageddo/dns-proxy-server
     volumes:
       - $PWD:/app/src/github.com/mageddo/dns-proxy-server
